Fix CTS service and CTS client read operations
CTS service and client read operations are not standards compliant.
Add missing day of week field (DOW) to the CTS service and CTS client.

Adding DOW and setting values for DOW, millis, and reason fixes CTS service read.

The CTS client read worked before this commit, but it was not standards compliant.
    It worked because we don't use the DOW, millis, or reason fields in the
    dateTimeController when we set the RTC.