Refactor, document, and fix the Watchdog driver
Refactor and document the Watchdog driver to make it more readable.

Fix the configuration of behaviours that were not properly implemented
(but didn't cause problems since the correct value was eventually set in NRF_WDT->CONFIG).

Fix the wrong interpretation of the reset reasons caused by implicit conversions of int to bool.

Increase the watchdog timeout from 7 seconds to 9 seconds